fix compile

+++ ebpf/bpfwrapper/eventCallbacks.go
@@ -64,7 +64,7 @@ var (
 	eventAttributesSize = int(unsafe.Sizeof(structs.SocketDataEventAttr{}))
 )
 
-func min(a, b int) int {
+func min(a, b int64) int64 {
 	if a < b {
 		return a
 	}
